refactor: extract aerial and whiff stats accumulators
